Support quantity functions around obstacles
This way there is smooth transition between the free fluid and the fluid
adjacent to an obstacle. Typically, one wants to supply the same functions to
both the background gas grid and the obstacles to achieve this.

The order of arguments to the constructors has been adapted to match the order
in the IC binary file for sake of consistency across the API.