Update the readme to be specific
It is unlikely that jelly will support other hydrodynamics software in the
future, at least there are no concrete plans to do so.

Should this happen nonetheless, I won't reject such an development. The readme
may be updated at that point but currently I think it is better to reflect the
status quo.